# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "JellyRick 🚀",
    "description": "API to post comments about Rick & Morty universe.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "User List Route",
        "description": "route to get user data",
        "operationId": "user_list_route_api_v1_user_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "User Route With Parameters",
        "description": "route to post user data.",
        "operationId": "user_route_with_parameters_api_v1_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/limit-offset": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "User List Route",
        "description": "route to get user data",
        "operationId": "user_list_route_api_v1_user_limit_offset_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "default": 50
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/{username}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "User Unique Route",
        "description": "route to get user data",
        "operationId": "user_unique_route_api_v1_user__username__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            },
            "name": "username",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "User Delete Route",
        "description": "route to delete user data",
        "operationId": "user_delete_route_api_v1_user__username__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            },
            "name": "username",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "LimitOffsetPage_str_": {
        "title": "LimitOffsetPage[str]",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0,
            "type": "integer"
          }
        }
      },
      "Page_str_": {
        "title": "Page[str]",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0,
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "minimum": 1,
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "minimum": 1,
            "type": "integer"
          }
        }
      },
...
      "User": {
        "title": "User",
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "title": "Username",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "disabled": {
            "title": "Disabled",
            "type": "boolean"
          },
          "administrator": {
            "title": "Administrator",
            "type": "boolean"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "boolean"
          },
          "moderator": {
            "title": "Moderator",
            "type": "boolean"
          }
        },
        "example": {
          "username": "johndoe",
          "full_name": "John Doe",
          "email": "johndoe@example.com",
          "disabled": false,
          "administrator": false,
          "reviewer": false,
          "moderator": false
        }
      },
      "UserPost": {
        "title": "UserPost",
        "type": "object",
        "properties": {},
        "example": {
          "content": {
            "username": "johndoe",
            "full_name": "John Doe",
            "email": "johndoe@example.com",
            "disabled": false,
            "administrator": false,
            "reviewer": false,
            "moderator": false
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "me": "Read information about the current user.",
              "items": "Read items."
            },
            "tokenUrl": "/api/v1/login"
          }
        }
      }
    }
  }
}